backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
django-storages==1.14.4
django-celery-beat==2.7.0
django-redis==5.4.0
django-environ==0.11.2
django-extensions==3.2.3

psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0

gunicorn==23.0.0
Pillow==11.1.0
boto3==1.35.81
stripe==11.4.1

bleach==6.2.0
python-slugify==8.0.4
django-mptt==0.16.0

flake8==7.1.1
black==24.10.0
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

sentry-sdk[django]==2.19.2

drf-spectacular==0.28.0
